feat: StreamTextScreen UI 개발-youtubePlayer 연동, Lexical 데이터 출력 및 하이라이트 표시, youtubePlayer 모듈 - 0.1초 단위로 재생시간 호출 로직 수정

+++ core/youtube-player/src/main/java/kr/co/are/searchcocktail/core/youtubeplayer/bridge/WebViewYoutubePlayerBridge.kt
@@ -4,10 +4,10 @@ import android.webkit.JavascriptInterface
 import timber.log.Timber
 
 class WebViewYoutubePlayerBridge(
-    private val onUpdatedYoutubePlayTime: (Int) -> Unit
+    private val onUpdatedYoutubePlayTime: (Float) -> Unit
 ) {
     @JavascriptInterface
-    fun onUpdatedPlayTime(time:Int) {
+    fun onUpdatedPlayTime(time:Float) {
         onUpdatedYoutubePlayTime(time)
     }
 }
